requirements.txt
flask==3.0.2
youtube_transcript_api==0.6.2
google-generativeai==0.3.2
gunicorn==21.2.0 